pyproject.toml
[project]
name = "internship-partone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.8",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.38",
]
